+++ config/babylon.go
@@ -3,7 +3,7 @@ package config
 import (
 	"time"
 
-	bbncfg "github.com/babylonlabs-io/babylon/v3/client/config"
+	bbncfg "github.com/babylonlabs-io/babylon/v4/client/config"
 )
 
 type BBNConfig struct {