+++ CHANGELOG.md
@@ -37,10 +37,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#160](https://github.com/babylonlabs-io/covenant-emulator/pull/160) chore: bump babylon to `v4.0.0-rc.3`
+* [#159](https://github.com/babylonlabs-io/covenant-emulator/pull/159) Support M-of-N multisig btc delegation
+
 ### Bug Fixes
 
 * [#156](https://github.com/babylonlabs-io/covenant-emulator/pull/156) fix: overflow in parse hardened
-* [#159](https://github.com/babylonlabs-io/covenant-emulator/pull/159) Support M-of-N multisig btc delegation
 
 ## v0.16.0-rc.2